+++ go/cert_srv/internal/csconfig/testdata/csconfig.toml
@@ -3,6 +3,3 @@
   IssuerReissueLeadTime = "2d"
   ReissueRate = "12s"
   ReissueTimeout = "6s"
-  SciondPath = "/run/shm/sciond/sd1-ff00_0_110.sock"
-  SciondTimeout = "20s"
-  SciondRetryInterval = "3s"
